[+] lend modal perfect

+++ src/layouts/index.tsx
@@ -4,6 +4,7 @@ import { Box } from "@material-ui/core";
 
 import "../style/index.scss";
 import { CurrentAddressContext } from "../../src/hardhat/SymfoniContext";
+import { short } from "../utils";
 
 type LayoutProps = {
   children: React.ReactNode;
@@ -22,7 +23,7 @@ const Layout: React.FC<LayoutProps> = ({ children }) => {
               <h1 className="Header__title">
                 <div data-text="ReNFT">ReNFT</div>
               </h1>
-              <div className="Header__summary">{currentAddress}</div>
+              <div className="Header__summary">{short(currentAddress)}</div>
             </div>
           </div>
         </div>
